+++ PaypalServerSdk.Standard/Controllers/PaymentsController.cs
@@ -64,6 +64,7 @@ namespace PaypalServerSdk.Standard.Controllers
                       .Header(_header => _header.Setup("PayPal-Auth-Assertion", input.PaypalAuthAssertion))))
               .ResponseHandler(_responseHandler => _responseHandler
                   .ErrorCase("401", CreateErrorCase("Authentication failed due to missing authorization header, or invalid authentication credentials.", (_reason, _context) => new ErrorException(_reason, _context)))
+                  .ErrorCase("403", CreateErrorCase("The request failed because the caller has insufficient permissions.", (_reason, _context) => new ErrorException(_reason, _context)))
                   .ErrorCase("404", CreateErrorCase("The request failed because the resource does not exist.", (_reason, _context) => new ErrorException(_reason, _context)))
                   .ErrorCase("500", CreateErrorCase("The request failed because an internal server error occurred.", (_reason, _context) => new ApiException(_reason, _context)))
                   .ErrorCase("0", CreateErrorCase("The error response.", (_reason, _context) => new ErrorException(_reason, _context))))
@@ -111,7 +112,7 @@ namespace PaypalServerSdk.Standard.Controllers
               .ExecuteAsync(cancellationToken).ConfigureAwait(false);
 
         /// <summary>
-        /// Reauthorizes an authorized PayPal account payment, by ID. To ensure that funds are still available, reauthorize a payment after its initial three-day honor period expires. Within the 29-day authorization period, you can issue multiple re-authorizations after the honor period expires. If 30 days have transpired since the date of the original authorization, you must create an authorized payment instead of reauthorizing the original authorized payment. A reauthorized payment itself has a new honor period of three days. You can reauthorize an authorized payment from 4 to 29 days after the 3-day honor period. The allowed amount depends on context and geography, for example in US it is up to 115% of the original authorized amount, not to exceed an increase of $75 USD. Supports only the `amount` request parameter. Note: This request is currently not supported for Partner use cases.
+        /// Reauthorizes an authorized PayPal account payment, by ID. To ensure that funds are still available, reauthorize a payment after its initial three-day honor period expires. Within the 29-day authorization period, you can issue multiple re-authorizations after the honor period expires. If 30 days have transpired since the date of the original authorization, you must create an authorized payment instead of reauthorizing the original authorized payment. A reauthorized payment itself has a new honor period of three days. You can reauthorize an authorized payment from 4 to 29 days after the 3-day honor period. The allowed amount depends on context and geography, for example in US it is up to 115% of the original authorized amount, not to exceed an increase of $75 USD. Supports only the `amount` request parameter.
         /// </summary>
         /// <param name="input">Object containing request parameters.</param>
         /// <returns>Returns the ApiResponse of Models.PaymentAuthorization response from the API call.</returns>
@@ -120,7 +121,7 @@ namespace PaypalServerSdk.Standard.Controllers
             => CoreHelper.RunTask(ReauthorizePaymentAsync(input));
 
         /// <summary>
-        /// Reauthorizes an authorized PayPal account payment, by ID. To ensure that funds are still available, reauthorize a payment after its initial three-day honor period expires. Within the 29-day authorization period, you can issue multiple re-authorizations after the honor period expires. If 30 days have transpired since the date of the original authorization, you must create an authorized payment instead of reauthorizing the original authorized payment. A reauthorized payment itself has a new honor period of three days. You can reauthorize an authorized payment from 4 to 29 days after the 3-day honor period. The allowed amount depends on context and geography, for example in US it is up to 115% of the original authorized amount, not to exceed an increase of $75 USD. Supports only the `amount` request parameter. Note: This request is currently not supported for Partner use cases.
+        /// Reauthorizes an authorized PayPal account payment, by ID. To ensure that funds are still available, reauthorize a payment after its initial three-day honor period expires. Within the 29-day authorization period, you can issue multiple re-authorizations after the honor period expires. If 30 days have transpired since the date of the original authorization, you must create an authorized payment instead of reauthorizing the original authorized payment. A reauthorized payment itself has a new honor period of three days. You can reauthorize an authorized payment from 4 to 29 days after the 3-day honor period. The allowed amount depends on context and geography, for example in US it is up to 115% of the original authorized amount, not to exceed an increase of $75 USD. Supports only the `amount` request parameter.
         /// </summary>
         /// <param name="input">Object containing request parameters.</param>
         /// <param name="cancellationToken"> cancellationToken. </param>
@@ -142,6 +143,7 @@ namespace PaypalServerSdk.Standard.Controllers
               .ResponseHandler(_responseHandler => _responseHandler
                   .ErrorCase("400", CreateErrorCase("The request failed because it is not well-formed or is syntactically incorrect or violates schema.", (_reason, _context) => new ErrorException(_reason, _context)))
                   .ErrorCase("401", CreateErrorCase("Authentication failed due to missing authorization header, or invalid authentication credentials.", (_reason, _context) => new ErrorException(_reason, _context)))
+                  .ErrorCase("403", CreateErrorCase("The request failed because the caller has insufficient permissions.", (_reason, _context) => new ErrorException(_reason, _context)))
                   .ErrorCase("404", CreateErrorCase("The request failed because the resource does not exist.", (_reason, _context) => new ErrorException(_reason, _context)))
                   .ErrorCase("422", CreateErrorCase("The request failed because it either is semantically incorrect or failed business validation.", (_reason, _context) => new ErrorException(_reason, _context)))
                   .ErrorCase("500", CreateErrorCase("The request failed because an internal server error occurred.", (_reason, _context) => new ApiException(_reason, _context)))