+++ PaypalServerSdk.Standard/Models/CaptureRequest.cs
@@ -33,45 +33,39 @@ namespace PaypalServerSdk.Standard.Models
         /// <summary>
         /// Initializes a new instance of the <see cref="CaptureRequest"/> class.
         /// </summary>
-        /// <param name="invoiceId">invoice_id.</param>
-        /// <param name="noteToPayer">note_to_payer.</param>
         /// <param name="amount">amount.</param>
+        /// <param name="invoiceId">invoice_id.</param>
         /// <param name="finalCapture">final_capture.</param>
         /// <param name="paymentInstruction">payment_instruction.</param>
+        /// <param name="noteToPayer">note_to_payer.</param>
         /// <param name="softDescriptor">soft_descriptor.</param>
         public CaptureRequest(
-            string invoiceId = null,
-            string noteToPayer = null,
             Models.Money amount = null,
+            string invoiceId = null,
             bool? finalCapture = false,
             Models.CapturePaymentInstruction paymentInstruction = null,
+            string noteToPayer = null,
             string softDescriptor = null)
         {
-            this.InvoiceId = invoiceId;
-            this.NoteToPayer = noteToPayer;
             this.Amount = amount;
+            this.InvoiceId = invoiceId;
             this.FinalCapture = finalCapture;
             this.PaymentInstruction = paymentInstruction;
+            this.NoteToPayer = noteToPayer;
             this.SoftDescriptor = softDescriptor;
         }
-
-        /// <summary>
-        /// The API caller-provided external invoice number for this order. Appears in both the payer's transaction history and the emails that the payer receives.
-        /// </summary>
-        [JsonProperty("invoice_id", NullValueHandling = NullValueHandling.Ignore)]
-        public string InvoiceId { get; set; }
-
-        /// <summary>
-        /// An informational note about this settlement. Appears in both the payer's transaction history and the emails that the payer receives.
-        /// </summary>
-        [JsonProperty("note_to_payer", NullValueHandling = NullValueHandling.Ignore)]
-        public string NoteToPayer { get; set; }
 
         /// <summary>
         /// The currency and amount for a financial transaction, such as a balance or payment due.
         /// </summary>
         [JsonProperty("amount", NullValueHandling = NullValueHandling.Ignore)]
         public Models.Money Amount { get; set; }
+
+        /// <summary>
+        /// The API caller-provided external invoice number for this order. Appears in both the payer's transaction history and the emails that the payer receives.
+        /// </summary>
+        [JsonProperty("invoice_id", NullValueHandling = NullValueHandling.Ignore)]
+        public string InvoiceId { get; set; }
 
         /// <summary>
         /// Indicates whether you can make additional captures against the authorized payment. Set to `true` if you do not intend to capture additional payments against the authorization. Set to `false` if you intend to capture additional payments against the authorization.
@@ -84,6 +78,12 @@ namespace PaypalServerSdk.Standard.Models
         /// </summary>
         [JsonProperty("payment_instruction", NullValueHandling = NullValueHandling.Ignore)]
         public Models.CapturePaymentInstruction PaymentInstruction { get; set; }
+
+        /// <summary>
+        /// An informational note about this settlement. Appears in both the payer's transaction history and the emails that the payer receives.
+        /// </summary>
+        [JsonProperty("note_to_payer", NullValueHandling = NullValueHandling.Ignore)]
+        public string NoteToPayer { get; set; }
 
         /// <summary>
         /// The payment descriptor on the payer's account statement.
@@ -106,16 +106,16 @@ namespace PaypalServerSdk.Standard.Models
             if (ReferenceEquals(this, obj)) return true;
 
             return obj is CaptureRequest other &&
-                (this.InvoiceId == null && other.InvoiceId == null ||
-                 this.InvoiceId?.Equals(other.InvoiceId) == true) &&
-                (this.NoteToPayer == null && other.NoteToPayer == null ||
-                 this.NoteToPayer?.Equals(other.NoteToPayer) == true) &&
                 (this.Amount == null && other.Amount == null ||
                  this.Amount?.Equals(other.Amount) == true) &&
+                (this.InvoiceId == null && other.InvoiceId == null ||
+                 this.InvoiceId?.Equals(other.InvoiceId) == true) &&
                 (this.FinalCapture == null && other.FinalCapture == null ||
                  this.FinalCapture?.Equals(other.FinalCapture) == true) &&
                 (this.PaymentInstruction == null && other.PaymentInstruction == null ||
                  this.PaymentInstruction?.Equals(other.PaymentInstruction) == true) &&
+                (this.NoteToPayer == null && other.NoteToPayer == null ||
+                 this.NoteToPayer?.Equals(other.NoteToPayer) == true) &&
                 (this.SoftDescriptor == null && other.SoftDescriptor == null ||
                  this.SoftDescriptor?.Equals(other.SoftDescriptor) == true);
         }
@@ -126,11 +126,11 @@ namespace PaypalServerSdk.Standard.Models
         /// <param name="toStringOutput">List of strings.</param>
         protected void ToString(List<string> toStringOutput)
         {
-            toStringOutput.Add($"InvoiceId = {this.InvoiceId ?? "null"}");
-            toStringOutput.Add($"NoteToPayer = {this.NoteToPayer ?? "null"}");
             toStringOutput.Add($"Amount = {(this.Amount == null ? "null" : this.Amount.ToString())}");
+            toStringOutput.Add($"InvoiceId = {this.InvoiceId ?? "null"}");
             toStringOutput.Add($"FinalCapture = {(this.FinalCapture == null ? "null" : this.FinalCapture.ToString())}");
             toStringOutput.Add($"PaymentInstruction = {(this.PaymentInstruction == null ? "null" : this.PaymentInstruction.ToString())}");
+            toStringOutput.Add($"NoteToPayer = {this.NoteToPayer ?? "null"}");
             toStringOutput.Add($"SoftDescriptor = {this.SoftDescriptor ?? "null"}");
         }
     }