+++ PaypalServerSdk.Standard/Models/CardVerificationDetails.cs
@@ -71,13 +71,13 @@ namespace PaypalServerSdk.Standard.Models
         public string Date { get; set; }
 
         /// <summary>
-        /// The card network or brand. Applies to credit, debit, gift, and payment cards.
+        /// DEPRECATED. This field is DEPRECATED. Please find the network data in the 'network' field under the 'network_transaction_reference' object instead of the 'verification' object.
         /// </summary>
         [JsonProperty("network", NullValueHandling = NullValueHandling.Ignore)]
         public Models.CardBrand? Network { get; set; }
 
         /// <summary>
-        /// The date and time, in [Internet date and time format](https://tools.ietf.org/html/rfc3339#section-5.6). Seconds are required while fractional seconds are optional. Note: The regular expression provides guidance but does not reject all invalid dates.
+        /// DEPRECATED. This field is DEPRECATED. Please find the time data in the 'time' field under the 'network_transaction_reference' object instead of the 'verification' object.
         /// </summary>
         [JsonProperty("time", NullValueHandling = NullValueHandling.Ignore)]
         public string Time { get; set; }